Tri des galeries Tri de l'index Administration des tris
Refonte de la structure html et suppression de toutes les tables

Début des galeries récursives =>  tout l'affichage est cassé !!!

Amélirations pour les commentaire


git-svn-id: file:///home/nicolas/svn/perso/luxbum/trunk@68 795f2d49-7102-0410-9315-962b31f5c57f

+++ _fonctions/class/luxbumimage.class.php
@@ -1,7 +1,9 @@
-<?php 
+<?php
 
 include (LIB_DIR.'exifer/exif.php');
 include_once (FONCTIONS_DIR.'mysql.inc.php');
+
+define ('NOT_SET', 'Not Set');
 
 //==============================================================================
 // Classe luxBumImage : Fonctions pour les générations de miniatures
@@ -18,16 +20,22 @@ class luxBumImage extends luxBum
    var $previewDir;
 
 
-   var $description;
-   var $date;
+   var $description = NULL;
+   var $date = NULL;
 
-   var $thumbToolkit;
-   var $previewToolkit;
+   var $thumbToolkit = NULL;
+   var $previewToolkit = NULL;
    
    var $previewImagePath;
+   
+   var $sortPosition = 0;
 
 
-
+   /**
+    * Constructeur par défaut
+    * @param String $dir le nom de la galerie
+    * @param String $img le nom de l'image
+    */
    function luxBumImage ($dir, $img) {
       $this->dir = $dir;
       $this->img = $img;
@@ -37,43 +45,84 @@ class luxBumImage extends luxBum
       $this->previewImagePath = $this->getPreviewImage ($this->dir, $this->img);
    }
    
+   /**
+    * Retourne le chemin de l'image d'aperçu
+    * @return String Chemin de l'image d'aperçu
+    */
    function getPreviewImagePath () {
       return $this->previewImagePath;
    }
 
+   /**
+    * Retourne le dossier de l'image
+    * @return String Dossier de l'image
+    */
    function getImageDir () {
       return $this->dir;
    }
 
+   /**
+    * Retourne le nom de l'image
+    * @return String Nom de l'image
+    */
    function getImageName () {
       return $this->img;
    }
 
+   /**
+    * Retourne le chemin complet de l'image
+    * @return String Chemin complet de l'image
+    */
    function getImagePath () {
       return $this->getImage ($this->dir, $this->img);
    }
 
+   /**
+    * Retourne la description de l'image
+    * @return String Description de l'image
+    */
    function getDescription () {
       return $this->description;
    }
 
+   /**
+    * Retourne la date de l'image
+    * @return String Date de l'image
+    */
    function getDate () {
       return $this->date;
    }
 
+   /**
+    * Affecte la description de l'image
+    * @param String $description Description de l'image
+    */
    function setDescription ($description) {
       $this->description = $description;
    }
 
+   /**
+    * Affecte la date de l'image
+    * @param String $date Date de l'image
+    */
    function setDate ($date) {
       $this->date = $date;
    }
 
+   /**
+    * Affecte la date et la description de l'image
+    * @param String $description Description de l'image
+    * @param String $date Date de l'image
+    */
    function setAllDescription ($description, $date) {
       $this->setDescription ($description);
       $this->setDate ($date);
    }
 
+   /**
+    * Retourne true/false si la date et la description sont vide
+    * @return Boolean true/false si la date et la description sont vide
+    */
    function issetDescription () {
       if ($this->description == '' && $this->date == '') {
          return false;
@@ -81,10 +130,18 @@ class luxBumImage extends luxBum
       return true;
    }
 
+   /**
+    * Retourne la taille en octets de l'image
+    * @return int Taille en octets de l'image
+    */
    function getSize () {
       return filesize ($this->getImagePath ());
    }
    
+   /**
+    * Retourne la date et la description sous un format affichable
+    * @return String Date et descrition sous format affichable
+    */
    function getDateDesc () {
       $dateDesc = '&nbsp;';
       
@@ -96,22 +153,54 @@ class luxBumImage extends luxBum
          $dateDesc = 'Le '.strftime (DATE_FORMAT,  $timeStamp);
     
          // date + description
-         if ($this -> getDescription () != '' && $this -> getDescription () != "\n") {
+         if ($this -> getDescription () != '') {
             $dateDesc .= ' - '. ucfirst ($this -> getDescription ());
          }
       }
    
       // Que description
-      else if ($this -> getDescription () != '' && $this -> getDescription () != "\n") {
+      else if ($this -> getDescription () != '') {
          $dateDesc = ucfirst ($this -> getDescription ());
       }
       return $dateDesc;
+   }
+   
+   /**
+    * Retourne le type mime de l'image
+    * @return Type mime de l'image
+    */
+   function getTypeMime () {
+      if ($this->thumbToolkit == null) {
+         return '';
+      }
+      return $this->thumbToolkit->getTypeMime();
+   }
+   
+   /**
+    * Affecte l'ordre manuel de tri
+    * @param int $sortOrder Ordre de tri
+    */
+   function setSortPosition ($sortPosition) {
+      $this->sortPosition = $sortPosition;
+   }
+   
+   /**
+    * Retourne l'ordre manuel de tri
+    * @return int Ordre manuel de tri
+    */
+   function getSortPosition () {
+      return $this->sortPosition;
    }
 
    /**-----------------------------------------------------------------------**/
    /** Fonctions des descriptions d'images */
    /**-----------------------------------------------------------------------**/
 
+   /**
+    * Recherche et affecte la date/description de l'image dans le fichier des
+    * descriptions.
+    * @return boolean true/false Date/description trouvés ou non
+    */
    function findDescription () {
 
       // Une description est déjà rentrée, pas besoin de chercher !
@@ -126,7 +215,7 @@ class luxBumImage extends luxBum
       if (is_file ($this->getDirPath ($this->getImageDir()).DESCRIPTION_FILE)) {
          $fd = fopen ($this->getDirPath ($this->getImageDir()).DESCRIPTION_FILE, 'r+');
          while (!$trouve && $line = fgets ($fd)) {
-            if ( ereg ('^.*\|.*\|.*$', $line)) {
+            if (ereg ('^.*\|.*\|.*$', $line)) {
                $tab = explode ('|', $line, 2);
                $desc[$tab[0]] = $tab[1];
                if ($tab[0] == $this->getImageName ()) {
@@ -151,8 +240,40 @@ class luxBumImage extends luxBum
    /**-----------------------------------------------------------------------**/
    /** Fonctions pour créer les thumbs / preview */
    /**-----------------------------------------------------------------------**/
+   
+   /**
+    * Retourne le lien de la vignette de l'image vers le script qui génére
+    * l'image
+    * @return Lien de la vignette de l'image vers le script de génération
+    */
+   function getThumbLink () {
+      if (USE_REWRITE == 'on') {
+         $prefix = 'image/';
+      }
+      else {
+         $prefix = 'image.php?';
+      }
+      return $prefix.THUMB_DIR.'-'.$this->dir.'-'.$this->img;
+   }
+   
+   /**
+    * Retourne le lien de l'aperçu de l'image vers le script qui génére l'image
+    * @return Lien de l'aperçu de l'image vers le script de génération
+    */
+   function getPreviewLink () {
+      if (USE_REWRITE == 'on') {
+         $prefix = 'image/';
+      }
+      else {
+         $prefix = 'image.php?';
+      }
+      return $prefix.PREVIEW_DIR.'-'.$this->dir.'-'.$this->img;
+   }
 
-
+   /**
+    * Génére la vignette de l'image et retourne le chemin vers l'image générée
+    * @return String Chemin vers la vignette générée.
+    */
    function getAsThumb ($dst_w = 85, $dst_h = 85) {
       $this->thumbToolkit = new imagetoolkit ($this->getImagePath ());
       $this->thumbToolkit->setDestSize ($dst_w, $dst_h);
@@ -162,51 +283,68 @@ class luxBumImage extends luxBum
          files::createDir ($this->thumbDir);
          $this->thumbToolkit->createThumb ($final);
       }
-      unset ($this->thumbToolkit);
       return $final;
    }
    
-   function needPreview ($dst_w = 650, $dst_h = 485) {
-      $this->previewToolkit = new imagetoolkit ($this->getImagePath ());
-      $this->previewToolkit->setDestSize ($dst_w, $dst_h);
-
-      // Pas de génération de preview
-      if ($this->getSize () < MIN_SIZE_FOR_PREVIEW * 1024) {
+   /**
+    * Cette fonction détermine si oui ou non il faut générer un aperçu.
+    * L'aperçu est généré seulement si la taille en octets de l'image est
+    * supérieur au seuil fixé.
+    * @access private
+    * @return boolean true/false Génére l'aperçu ou non
+    */
+   function _needPreview () {
+      if ($this->getSize() < MIN_SIZE_FOR_PREVIEW * 1024) {
          return false;
       }
-      
-      // Si image de départ plus petite, on ne redimentione pas la photo
-      if ($this->previewToolkit->destBiggerThanFrom()) {
-         return false;
-      }
-      
       return true;
    }
 
+   /**
+    * Génére l'aperçu de l'image et retourne le chemin vers l'image générée
+    * @return String Chemin vers l'aperçu généré.
+    */
    function getAsPreview ($dst_w = 650, $dst_h = 485) {
+      $this->previewToolkit = new imagetoolkit ($this->getImagePath ());
+      $this->previewToolkit->setDestSize ($dst_w, $dst_h);
       
-      if ($this->needPreview($dst_w, $dst_h) == false) {
+      // Si pas d'aperçu on retourne l'image originale
+      if ($this->_needPreview() == false) {
          return $this->getImagePath ();
       }
 
       // Génération de preview
-      //$final = $this->getPreviewImage ($this->dir, $this->img);
       if (!is_file ($this->previewImagePath)) {
          files::createDir ($this->previewDir);
          $this->previewToolkit->createThumb ($this->previewImagePath);
       }
-      unset ($this->thumbToolkit);
       return $this->previewImagePath;
    }
 
+   /**
+    * Retourne la chaine de taille de la vignette pour la balise &gt;img&lt;
+    * @return String Taille de la vignette pour la balise &gt;img&lt;
+    */
    function getThumbResizeSize () {
-      return 'width="'.($this->thumbToolkit->getImageDestWidth ())
-         .'" height="'.($this->thumbToolkit->getImageDestHeight ()).'"';
+      if ($this->thumbToolkit == null) {
+         return '';
+      }
+      return sprintf ('width="%s" height="%s"',
+                      $this->thumbToolkit->getImageDestWidth(),
+                      $this->thumbToolkit->getImageDestHeight());
    }
 
+   /**
+    * Retourne la chaine de taille de l'aperçu pour la balise &gt;img&lt;
+    * @return String Taille de l'aperçu pour la balise &gt;img&lt;
+    */
    function getPreviewResizeSize () {
-      return 'width="'.($this->previewToolkit->imageDestWidth)
-         .'" height="'.($this->previewToolkit->imageDestHeight).'"';
+      if ($this->previewToolkit == null) {
+         return '';
+      }
+      return sprintf ('width="%s" height="%s"',
+                      $this->previewToolkit->getImageDestWidth(),
+                      $this->previewToolkit->getImageDestHeight());
    }
 
 
@@ -215,7 +353,8 @@ class luxBumImage extends luxBum
    /**-----------------------------------------------------------------------**/
 
    /**
-    * 
+    * Supprime la photo ainsi que tout son cache et les commentaires associés.
+    * @return Boolean
     */
    function delete () {
       $this->clearCache ();
@@ -225,7 +364,7 @@ class luxBumImage extends luxBum
 
 
    /**
-    * 
+    * Supprime le cache de l'image
     */
    function clearCache () {
       $this->clearThumbCache ();
@@ -233,24 +372,24 @@ class luxBumImage extends luxBum
    }
 
    /**
-    *
+    * Supprime le cache des aperçus
     */
    function clearThumbCache () {
 
-      $fd = opendir ($this->thumbDir);
-
-      while ($current_file = readdir ($fd)) {
-         if ($current_file[0] != '.' 
-             && !is_dir ($this->thumbDir.$current_file) 
-             && eregi ('^.*(' . $this->img . ')$', $current_file)){
-            files::deleteFile ($this->thumbDir.$current_file);
+      if ($fd = opendir ($this->thumbDir)) {
+         while ($current_file = readdir ($fd)) {
+            if ($current_file[0] != '.' 
+                && !is_dir ($this->thumbDir.$current_file) 
+                && eregi ('^.*(' . $this->img . ')$', $current_file)){
+               files::deleteFile ($this->thumbDir.$current_file);
+            }
          }
+         closedir ($fd);
       }
-      closedir ($fd);
    }
 
    /**
-    *
+    * Supprime le cache des vignettes
     */
    function clearPreviewCache () {
       files::deleteFile ($this->previewDir . $this->img);
@@ -260,10 +399,17 @@ class luxBumImage extends luxBum
    /**-----------------------------------------------------------------------**/
    /** Fonctions pour les commentaires */
    /**-----------------------------------------------------------------------**/
+   /**
+    * Retourne le nombre de commentaires actifs de la photos
+    * @return nombre de commentaires actifs
+    */
    function getNbComment () {
       global $mysql;
-      $query = "SELECT count(*) FROM ".DBL_PREFIX."commentaire "
-         ."WHERE galerie_comment='".($this->dir)."' AND photo_comment='".($this->img)."'";
+      $query = sprintf ("SELECT count(*) FROM ".DBL_PREFIX."commentaire "
+         ."WHERE galerie_comment=%s AND photo_comment=%s AND pub_comment=%s",
+         $mysql->escapeString($this->dir),
+         $mysql->escapeString($this->img),
+         $mysql->escapeSet(1));
       return $mysql->DbCount ($query);
    }
 
@@ -271,11 +417,10 @@ class luxBumImage extends luxBum
    /** Fonctions d'informations exif */
    /**-----------------------------------------------------------------------**/
    var $exifResult;
-   var $notSet = 'Not Set';
 
 
    /**
-    *
+    * @access private
     */
    function &reduceExif ($exifvalue) {
       $vals = split ("/",$exifvalue);
@@ -293,7 +438,7 @@ class luxBumImage extends luxBum
    }
 
    /**
-    *
+    * @access private
     */
    function pullout ($str){
       $str = stripslashes($str);
@@ -302,7 +447,7 @@ class luxBumImage extends luxBum
    }
 
    /**
-    *
+    * Initialise les informations EXIF de la photo
     */
    function exifInit () {
       $verbose = 0;
@@ -310,7 +455,8 @@ class luxBumImage extends luxBum
    }
 
    /**
-    *
+    * Retourne ok si les informations EXIF existent
+    * @return boolean Infomations EXIF existent
     */
    function exifExists () {
       if (array_key_exists ('SubIFD', $this->exifResult) && 
@@ -321,46 +467,52 @@ class luxBumImage extends luxBum
    }
 
    /**
-    *
+    * Retourne la valeur ISO
+    * @return String Valeur ISO
     */
    function getExifISO () {
       if (array_key_exists ('ISOSpeedRatings', $this->exifResult['SubIFD'])) {
          return $this->pullout ($this->exifResult['SubIFD']['ISOSpeedRatings']);
       }
-      return $this->notSet;
+      return NOT_SET;
    }
+   
    /**
-    *
+    * Retourne le modèle de l'appareil photo
+    * @return Modèle de l'appareil photo
     */
    function getExifCameraModel () {
       if (array_key_exists ('Model', $this->exifResult['IFD0'])) {
          return trim ($this->exifResult['IFD0']['Model']);
       }
-      return $this->notSet;
+      return NOT_SET;
    }
 
    /**
-    *
+    * Retourne la marque de l'appareil photo
+    * @return String Marque de l'appareil photo
     */
    function getExifCameraMaker () {
       if (array_key_exists ('Make', $this->exifResult['IFD0'])) {
          return trim ($this->exifResult['IFD0']['Make']);
       }
-      return $this->notSet;
+      return NOT_SET;
    }
 
    /**
-    *
+    * Retourne la distance focale
+    * @return String Distance focale
     */
    function getExifFocalLength () {
       if (array_key_exists ('FocalLength', $this->exifResult['SubIFD'])) {
          return $this->exifResult['SubIFD']['FocalLength'];
       }
-      return $this->notSet;
+      return NOT_SET;
    }
 
    /**
-    *
+    * Retourne si le flash c'est déclenché ou non
+    * @return Boolean Flash déclenché ou non
     */
    function getExifFlash () {
       if (array_key_exists ('Flash', $this->exifResult['SubIFD'])) {
@@ -370,31 +522,34 @@ class luxBumImage extends luxBum
          }
          return $flash;
       }
-      return $this->notSet;
+      return NOT_SET;
    }
 
    /**
-    *
+    * Retourne la date à laquelle la photo a été prise
+    * @return Date de prise de la photo
     */
    function getExifCaptureDate () {
       if (array_key_exists ('DateTimeOriginal', $this->exifResult['SubIFD'])) {
          return $this->exifResult['SubIFD']['DateTimeOriginal'];
       }
-      return $this->notSet;
+      return NOT_SET;
    }
 
    /**
-    *
+    * Retourne le temps d'ouverture
+    * @return String Temps d'ouverture
     */
    function getExifAperture () {
       if (array_key_exists ('FNumber', $this->exifResult['SubIFD'])) {
          return $this->exifResult['SubIFD']['FNumber'];
       }
-      return $this->notSet;
+      return NOT_SET;
    }
 
    /**
-    *
+    * Retourne le temps d'exposition
+    * @return String Temps d'exposition
     */
    function getExifExposureTime () {
       if (array_key_exists ('ExposureTime', $this->exifResult['SubIFD'])) {
@@ -405,7 +560,7 @@ class luxBumImage extends luxBum
          }
          return $exposure;
       }
-      return $this->notSet;
+      return NOT_SET;
    }
 }
